'feat. 필수 기능 구현'

+++ src/main/java/com/sparta/currency_user/dto/CurrencyRequestDto.java
@@ -3,11 +3,13 @@ package com.sparta.currency_user.dto;
 import com.sparta.currency_user.entity.Currency;
 import lombok.Getter;
 
+import java.math.BigDecimal;
+
 
 @Getter
 public class CurrencyRequestDto {
     private String currencyName;
-    private Double exchangeRate;
+    private BigDecimal exchangeRate;
     private String symbol;
 
     public Currency toEntity() {
